Make CHIP-8 display work!

+++ src/EmulationScreenWidget.cpp
@@ -20,11 +20,11 @@ void EmulationScreenWidget::paintEvent(QPaintEvent * event) {
     const uint16_t& xRes = Screen::xRes_;
     const uint16_t& yRes = Screen::yRes_;
 
-    uint16_t pixelWidth = width() / xRes;
-    uint16_t pixelHeight = height() / yRes;
+    uint32_t pixelWidth = width() / xRes;
+    uint32_t pixelHeight = height() / yRes;
 
-    uint16_t xOffset = width() % xRes / 2;
-    uint16_t yOffset = height() % yRes / 2;
+    uint32_t xOffset = width() % xRes / 2;
+    uint32_t yOffset = height() % yRes / 2;
     
     for(uint16_t y = 0; y < yRes; ++y) {
         for(uint16_t x = 0; x < xRes; ++x) {
